Modularise index pages.
